Move OpenMP workspace setup inside the loop for easier error handling.
This avoids having to catch exceptions separately during workspace setup.
Doing so would be difficult as the omp-for loop itself would also need to move
inside the setup's try{}, and I'm not sure how the compiler would handle that.

Alternatively, we could put the workspace on the heap so that each thread can
just test whether its workspace was properly constructed (based on a non-NULL
pointer) before running the range inside the omp-for loop. But that's more
complicated for the user and probably takes a performance hit.

So, the easiest solution is to move the workspace construction inside the
omp-for loop. This is not ideal if the OpenMP runtime maps multiple workers to
a single thread, in which case redundant workspaces are constructed; but that
seems rare enough, usually OpenMP just uses the requested number of threads.

Also did a minor fix from incomplete variable renaming in the OpenMP code.