Removed the setup function from parallelize(), for simplicity.
Some practical experience with the workspace setup function indicates that it
introduces too much complexity. The allocation of data structures is separated
from their use, making it difficult to reason about the code. This is
especially bad when the allocation logic is split up between the setup function
and a non-default constructor of the workspace class. Moving the setup back
into run_task_range() makes life much simpler for the caller.

This change also reduces the potential for bugs when re-using data structures.
For example, we need to always make sure that any vectors are cleared before
push_back() or insert(), lest they contain some baggage from previous uses.
Failing to do so will introduce bugs that are very rare and difficult to test.
This is because such bugs won't show up with the default parallelization
schemes, which use a new workspace for every run_task_range() invocation.  To
check that re-use is done properly, each downstream library would have to
create separate executables with a SUBPAR_CUSTOM_PARALLEL override that forces
re-use of workspaces. This is tedious so we remove it to reduce the bug risk.

On a related note, the setup function doesn't confer much benefit for the
problems it introduces. The default schemes don't re-use the workspaces at all.
Custom schemes _might_ be able to take advantage of this but this will be
marginal if the task ranges are coarse-grained... which they probably should
be, as the run_task_range() might be doing arbitrarily expensive range-specific
setup, so a custom scheme wouldn't want to invoke that function too many times.