Added sanitize_num_workers() to clean up the number of workers.
This is not for internal use, but rather, for applications to call so
that they can better specify the number of workers (e.g., for their own
pre-allocations) before calling parallelize_range().

We also improve parallelize_range() so that we avoid an unnecessary
division if the number of workers is greater than the number of threads.