Added test function to easily check for correct use of parallelize_range().
The aim is to allow developers to easily check the assumptions around
the use of the worker ID (i.e., 'w') in downstream library code. The
default scheme has a nice and simple 1:1 mapping of calls to workers,
but developers should not assume that this is true for custom schemes
where run_task_range() might be called multiple times per worker.